- adds token search for api
- adds rebuild-cache command
- now returns total token count on api
- fixes various minor bugs

+++ src/run.js
@@ -4,6 +4,7 @@ import { find as findConfig } from './lib/config.js'
 import { load as loadConfig } from './lib/config.js'
 import { override as overrideConfig } from './lib/config.js'
 import startApp from './app/main.js'
+import rebuildCache from './cmd/rebuild-cache.js'
 
 
 
@@ -21,12 +22,17 @@ log.info(`using config at "${configPath}"`)
 const baseConfig = loadConfig(configPath, true)
 const config = overrideConfig(baseConfig, args)
 
-log.info(`data directory is at "${config.data.dir}"`)
-log.info(`will start app now`)
+if(args._[0] === 'rebuild-cache'){
+	log.info(`rebuilding cache at "${config.data.dir}"`)
+	await rebuildCache({ config })
+}else{
+	log.info(`data directory is at "${config.data.dir}"`)
+	log.info(`will start app now`)
 
-const app = await startApp({ config })
+	const app = await startApp({ config })
 
-process.on('SIGINT', async () => {
-	await app.terminate()
-	process.exit(0)
-})
+	process.on('SIGINT', async () => {
+		await app.terminate()
+		process.exit(0)
+	})
+}
